pyproject.toml
[project]
name = "releez"
description = "CLI tool for helping to manage release processes."
readme = "README.md"
requires-python = ">=3.11"
version = "0.0.0"

dependencies = [
  "typer ==0.20.0",
  "GitPython >= 3.1.45, <4",
  "PyGithub >= 2.8.1, <3",
  "git-cliff >=2.11.0, <3",
  "semver >=3.0.1, <4",
]

[dependency-groups]
dev = [
  "pytest >=9.0.2, <10",
  "pytest-cov >=7.0.0, <8",
  "pytest-mock >=3.15.1, <4",
]

[project.scripts]
releez = "releez.cli:main"

[tool.ruff]
extend = ".codeguide/configs/ruff.toml"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv.build-backend]
module-name = "releez"
module-root = "src"

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"
